Make latest migration more stable against error

+++ db/migrate/20130917094727_add_lot_of_missing_indexes.rb
@@ -17,8 +17,9 @@ class AddLotOfMissingIndexes < ActiveRecord::Migration
      add_index :people, :id
 
      # remove old style category indexes
-     remove_index :listings, :name => "index_listings_on_category"
-     remove_index :listings, :name => "index_listings_on_share_type"
+     
+     remove_index :listings, :name => "index_listings_on_category" if index_name_exists?(:listings, "index_listings_on_category", false)
+     remove_index :listings, :name => "index_listings_on_share_type" if index_name_exists?(:listings, "index_listings_on_share_type", false)
 
      add_index :listings, :category_id
      add_index :listings, :share_type_id
